Add functionality to update the number of likes after a movie is liked
